LPS-117064 Move to the api module to be able to reuse
