+++ modules/apps/layout/layout-api/src/main/resources/com/liferay/layout/util/structure/common-styles.json
